Move news generation to presenter
